requirements.txt
streamlit>=1.35.0
transformers>=4.41.0
torch>=2.2.0
pandas>=2.2.0
python-dotenv>=1.0.1